refactor cooking service tests
